refactor: split CRUD APIs into domain modules
