+++ src/utils/api/index.ts
@@ -1,0 +1,3 @@
+export * from './vehicles';
+export * from './drivers';
+export * from './trips';
