+++ src/utils/storage.ts
@@ -1,43 +1,11 @@
-import { supabase } from './supabaseClient';
-import { isNetworkError, handleNetworkError } from './supabaseClient';
+import { supabase, isNetworkError, handleNetworkError } from './supabaseClient';
 import config from './config';
-import {
-  Trip,
-  TripFormData,
-  Vehicle,
-  Driver,
-  DriverSummary,
-  Destination,
-  Warehouse,
-  RouteAnalysis,
-  AIAlert
-} from '../types';
-import { uploadVehicleDocument } from './supabaseStorage';
+import { Trip, Destination, Warehouse, RouteAnalysis, AIAlert } from '../types';
 import { getCurrentUserId, withOwner } from './supaHelpers';
 import { loadGoogleMaps } from './googleMapsLoader';
-import { toast } from 'react-toastify';
-import { generateCSV, downloadCSV } from './csvParser';
 import { handleSupabaseError } from './errors';
-
-// Helper function to check if a value is LTT (Lifetime Tax)
-const isLTT = (value: string | null | undefined): boolean => {
-  if (!value || typeof value !== 'string') return false;
-  const upperValue = value.trim().toUpperCase();
-  return ['LTT', 'LIFETIME', 'LIFE TIME', 'LIFE TIME TAX'].includes(upperValue);
-};
-
-// Helper function to get future date (10 years from now)
-const getFutureDateForLTT = (): string => {
-  const futureDate = new Date();
-  futureDate.setFullYear(futureDate.getFullYear() + 10);
-  return futureDate.toISOString().split('T')[0]; // YYYY-MM-DD format
-};
-
-// Define allowed columns for vehicles table
-export const VEHICLE_COLS = 'id,registration_number,make,model,year,type,fuel_type,current_odometer,status,chassis_number,engine_number,owner_name,tyre_size,number_of_tyres,rc_expiry_date,rc_document_url,insurance_policy_number,insurer_name,insurance_start_date,insurance_expiry_date,insurance_premium_amount,insurance_document_url,fitness_certificate_number,fitness_issue_date,fitness_expiry_date,fitness_cost,fitness_document_url,tax_receipt_number,tax_amount,tax_period,tax_document_url,permit_number,permit_issuing_state,permit_type,permit_issue_date,permit_expiry_date,permit_cost,permit_document_url,puc_certificate_number,puc_issue_date,puc_expiry_date,puc_cost,puc_document_url,issuing_state,other_documents,registration_date,policy_number,tax_receipt_document,insurance_idv,tax_scope,remind_insurance,insurance_reminder_contact_id,insurance_reminder_days_before,remind_fitness,fitness_reminder_contact_id,fitness_reminder_days_before,remind_puc,puc_reminder_contact_id,puc_reminder_days_before,remind_tax,tax_reminder_contact_id,tax_reminder_days_before,remind_permit,permit_reminder_contact_id,permit_reminder_days_before,remind_service,service_reminder_contact_id,service_reminder_days_before,service_reminder_km,photo_url,financer,vehicle_class,color,cubic_capacity,cylinders,unladen_weight,seating_capacity,emission_norms,noc_details,national_permit_number,national_permit_upto,rc_status,vahan_last_fetched_at,other_info_documents,tax_paid_upto,tags,last_updated_at,added_by,service_interval_km,service_interval_days,created_by,primary_driver_id,created_at,updated_at';
-
-// Define allowed columns for drivers table
-export const DRIVER_COLS = 'id,name,license_number,contact_number,email,join_date,status,experience_years,primary_vehicle_id,driver_photo_url,license_doc_url,aadhar_doc_url,police_doc_url,bank_doc_url,address,last_updated_at,blood_group,dob,father_or_husband_name,gender,license_issue_date,other_documents,rto,rto_code,state,valid_from,vehicle_class,notes,medical_doc_url,added_by,created_by,created_at,updated_at,license_expiry_date';
+import { getVehicle, deleteVehicle } from './api';
+export * from './api';
 
 export async function getUserData() {
   try {
@@ -64,494 +32,6 @@ export async function getUserData() {
     throw error;
   }
 }
-
-// Vehicle CRUD operations
-export const getVehicles = async (): Promise<Vehicle[]> => {
-  try {
-    const { data: { user }, error: userError } = await supabase.auth.getUser();
-    
-    if (userError) {
-      if (isNetworkError(userError)) {
-        if (config.isDev) console.warn('Network error fetching user for vehicles, returning empty array');
-        return [];
-      }
-      handleSupabaseError('get user for vehicles', userError);
-      return [];
-    }
-    
-    if (!user) {
-      console.error('No user authenticated');
-      return [];
-    }
-
-    const { data, error } = await supabase
-      .from('vehicles')
-      .select(VEHICLE_COLS)
-      .eq('added_by', user.id)
-      .order('created_at', { ascending: false });
-
-    if (error) {
-      handleSupabaseError('fetch vehicles', error);
-      return [];
-    }
-
-    return data || [];
-  } catch (error) {
-    if (isNetworkError(error)) {
-      if (config.isDev) console.warn('Network error fetching user for vehicles, returning empty array');
-      return [];
-    }
-    handleSupabaseError('get user for vehicles', error);
-    return [];
-  }
-};
-
-export const getVehicle = async (id: string): Promise<Vehicle | null> => {
-  const { data, error } = await supabase
-    .from('vehicles')
-    .select(VEHICLE_COLS)
-    .eq('id', id)
-    .single();
-
-  if (error) {
-    handleSupabaseError('fetch vehicle', error);
-    return null;
-  }
-
-  return data;
-};
-
-export const createVehicle = async (vehicleData: Omit<Vehicle, 'id'>): Promise<Vehicle | null> => {
-  try {
-    const userId = await getCurrentUserId();
-    if (!userId) {
-      throw new Error('User not authenticated');
-    }
-
-    // Strip UI-only fields and prepare data
-    const { documents, selected, ...cleanData } = vehicleData as any; // ⚠️ Confirm field refactor here
-    
-    // Handle LTT (Lifetime Tax) conversion
-    // Replacing "LTT" with a valid future date (10 years from now) to match Supabase date format
-    if (isLTT(cleanData.tax_paid_upto)) {
-      cleanData.tax_paid_upto = getFutureDateForLTT();
-      // Save original LTT info in tax_scope for UI badge display
-      if (!cleanData.tax_scope || cleanData.tax_scope.trim() === '') {
-        cleanData.tax_scope = 'Lifetime Tax (LTT)';
-      }
-    }
-    
-    // Remove _file properties (client-side File objects, not database columns)
-    const payload = withOwner({
-      ...cleanData,
-      current_odometer: cleanData.current_odometer || 0,
-    }, userId);
-
-    // Strip file properties that shouldn't go to database
-    for (const k of Object.keys(payload)) {
-      if (k.endsWith('_file')) {
-        delete payload[k];
-      }
-    }
-
-    // Sanitize empty strings to null for database compatibility
-    for (const key in payload) {
-      if (typeof payload[key] === 'string' && payload[key].trim() === '') {
-        payload[key] = null;
-      }
-    }
-    const { data, error } = await supabase
-      .from('vehicles')
-      .insert(payload)
-      .select(VEHICLE_COLS)
-      .single();
-
-    if (error) {
-      handleSupabaseError('create vehicle', error);
-      throw error;
-    }
-
-    return data;
-  } catch (error) {
-    handleSupabaseError('create vehicle', error);
-    throw error;
-  }
-};
-
-export const updateVehicle = async (id: string, values: any): Promise<Vehicle | null> => {
-  try {
-    // Strip UI-only fields and undefined values
-    const {
-      documents, 
-      selected, // UI fields
-      stats, // Frontend-only field
-      odometer, 
-      odometer_km,
-      ...rest
-    } = values || {};
-
-    const payload: any = {
-      ...rest,
-      current_odometer: odometer_km ?? odometer ?? rest.current_odometer,
-    };
-
-    // Handle LTT (Lifetime Tax) conversion for updates
-    // Replacing "LTT" with a valid future date (10 years from now) to match Supabase date format
-    if (isLTT(payload.tax_paid_upto)) {
-      payload.tax_paid_upto = getFutureDateForLTT();
-      // Save original LTT info in tax_scope for UI badge display
-      if (!payload.tax_scope || payload.tax_scope.trim() === '') {
-        payload.tax_scope = 'Lifetime Tax (LTT)';
-      }
-    }
-
-    // Coerce current_odometer to number
-    if (payload.current_odometer != null) {
-      const n = Number(payload.current_odometer);
-      payload.current_odometer = Number.isFinite(n) ? n : null;
-    }
-
-    // Remove undefined keys
-    for (const k of Object.keys(payload)) {
-      if (payload[k] === undefined) {
-        delete payload[k];
-      }
-    }
-
-    // Remove _file properties (client-side File objects, not database columns)
-    for (const k of Object.keys(payload)) {
-      if (k.endsWith('_file')) {
-        delete payload[k];
-      }
-    }
-
-    // Update the vehicle
-    const { data, error } = await supabase
-      .from('vehicles')
-      .update(payload)
-      .eq('id', id)
-      .select(VEHICLE_COLS)
-      .single();
-
-    if (error) {
-      handleSupabaseError('update vehicle', error);
-      throw error;
-    }
-
-    // Handle document uploads safely
-    if (Array.isArray(documents)) {
-      for (const file of documents) {
-        // Guard against invalid files
-        if (!file || typeof file.name !== 'string') {
-          continue;
-        }
-
-        try {
-          const dotIndex = file.name.lastIndexOf('.');
-          const extension = dotIndex >= 0 ? file.name.slice(dotIndex + 1) : '';
-          
-          // Only proceed if we have a valid extension
-          if (extension) {
-            const filePath = await uploadVehicleDocument(file, id, 'general');
-            // Get existing other_documents array and add the new path
-            const existingDocs = data.other_documents || [];
-            const updatedDocs = [
-              ...existingDocs,
-              {
-                name: file.name,
-                file_path: filePath,
-                upload_date: new Date().toISOString()
-              }
-            ];
-            
-            // Update the vehicle with the new other_documents array
-            await supabase
-              .from('vehicles')
-              .update({ other_documents: updatedDocs })
-              .eq('id', id);
-          }
-        } catch (uploadError) {
-          handleSupabaseError('upload vehicle document', uploadError);
-          // Continue with other documents even if one fails
-        }
-      }
-    }
-
-    return data;
-  } catch (error) {
-    handleSupabaseError('update vehicle', error);
-    throw error;
-  }
-};
-
-export const deleteVehicle = async (id: string): Promise<boolean> => {
-  const { error } = await supabase
-    .from('vehicles')
-    .delete()
-    .eq('id', id);
-
-  if (error) {
-    handleSupabaseError('delete vehicle', error);
-    return false;
-  }
-
-  return true;
-};
-
-// Bulk archive vehicles
-export const bulkArchiveVehicles = async (vehicleIds: string[]): Promise<boolean> => {
-  try {
-    const { error } = await supabase
-      .from('vehicles')
-      .update({ status: 'archived' })
-      .in('id', vehicleIds);
-
-    if (error) {
-      handleSupabaseError('archive vehicles', error);
-      return false;
-    }
-
-    return true;
-  } catch (error) {
-    handleSupabaseError('archive vehicles', error);
-    return false;
-  }
-};
-
-// Export vehicle data to CSV
-export const exportVehicleData = async (): Promise<void> => {
-  try {
-    const vehicles = await getVehicles();
-    
-    // Define CSV headers
-    const headers = {
-      registration_number: 'Registration Number',
-      make: 'Make',
-      model: 'Model',
-      year: 'Year',
-      type: 'Type',
-      fuel_type: 'Fuel Type',
-      current_odometer: 'Current Odometer (km)',
-      status: 'Status',
-      chassis_number: 'Chassis Number',
-      engine_number: 'Engine Number',
-      owner_name: 'Owner Name',
-      created_at: 'Created Date',
-      updated_at: 'Updated Date'
-    };
-    
-    // Generate CSV
-    const csvData = await generateCSV(vehicles, headers);
-    
-    // Download CSV
-    const fileName = `vehicles_export_${new Date().toISOString().split('T')[0]}.csv`;
-    downloadCSV(fileName, csvData);
-  } catch (error) {
-    handleSupabaseError('export vehicle data', error);
-    throw error;
-  }
-};
-
-// Bulk unarchive vehicles
-export const bulkUnarchiveVehicles = async (vehicleIds: string[]): Promise<boolean> => {
-  try {
-    const { error } = await supabase
-      .from('vehicles')
-      .update({ status: 'active' })
-      .in('id', vehicleIds);
-
-    if (error) {
-      handleSupabaseError('unarchive vehicles', error);
-      return false;
-    }
-
-    return true;
-  } catch (error) {
-    handleSupabaseError('unarchive vehicles', error);
-    return false;
-  }
-};
-
-// Driver CRUD operations
-export const getDrivers = async (): Promise<Driver[]> => {
-  try {
-    const { data: { user }, error: userError } = await supabase.auth.getUser();
-    
-    if (userError) {
-      if (isNetworkError(userError)) {
-        if (config.isDev) console.warn('Network error fetching user for drivers, returning empty array');
-        return [];
-      }
-      handleSupabaseError('get user for drivers', userError);
-      return [];
-    }
-    
-    if (!user) {
-      console.error('No user authenticated');
-      return [];
-    }
-
-    const { data, error } = await supabase
-      .from('drivers')
-      .select(DRIVER_COLS)
-      .eq('added_by', user.id)
-      .order('created_at', { ascending: false });
-
-    if (error) {
-      handleSupabaseError('fetch drivers', error);
-      return [];
-    }
-
-    return data || [];
-  } catch (error) {
-    if (isNetworkError(error)) {
-      if (config.isDev) console.warn('Network error fetching user for drivers, returning empty array');
-      return [];
-    }
-    handleSupabaseError('get user for drivers', error);
-    return [];
-  }
-};
-
-export const getDriverSummaries = async (): Promise<DriverSummary[]> => {
-  try {
-    const { data: { user }, error: userError } = await supabase.auth.getUser();
-
-    if (userError) {
-      if (isNetworkError(userError)) {
-        if (config.isDev) console.warn('Network error fetching user for driver summaries, returning empty array');
-        return [];
-      }
-      handleSupabaseError('get user for driver summaries', userError);
-      return [];
-    }
-
-    if (!user) {
-      console.error('No user authenticated');
-      return [];
-    }
-
-    const { data, error } = await supabase
-      .from('drivers')
-      .select('id,name')
-      .eq('added_by', user.id)
-      .order('created_at', { ascending: false });
-
-    if (error) {
-      handleSupabaseError('fetch driver summaries', error);
-      return [];
-    }
-
-    return data || [];
-  } catch (error) {
-    if (isNetworkError(error)) {
-      if (config.isDev) console.warn('Network error fetching user for driver summaries, returning empty array');
-      return [];
-    }
-    handleSupabaseError('get user for driver summaries', error);
-    return [];
-  }
-};
-
-export const getDriver = async (id: string): Promise<Driver | null> => {
-  const { data, error } = await supabase
-    .from('drivers')
-    .select(DRIVER_COLS)
-    .eq('id', id)
-    .single();
-
-  if (error) {
-    handleSupabaseError('fetch driver', error);
-    return null;
-  }
-
-  return data;
-};
-
-export const createDriver = async (driverData: Omit<Driver, 'id'>): Promise<Driver | null> => {
-  try {
-    const userId = await getCurrentUserId(); // ⚠️ Confirm field refactor here
-    if (!userId) { // ⚠️ Confirm field refactor here
-      throw new Error('User not authenticated');
-    }
-
-    // Strip file objects and prepare payload
-    const { photo, aadhar_doc_file, medical_doc_file, license_doc_file, police_doc_file, ...cleanData } = driverData as any;
-    
-    const payload = withOwner(cleanData, userId);
-    
-    // Remove _file properties (client-side File objects, not database columns)
-    for (const k of Object.keys(payload)) {
-      if (k.endsWith('_file')) {
-        delete payload[k];
-      }
-    }
-
-    // Sanitize empty strings to null for database compatibility
-    for (const key in payload) {
-      if (typeof payload[key] === 'string' && payload[key].trim() === '') {
-        payload[key] = null;
-      }
-    }
-
-    const { data, error } = await supabase
-      .from('drivers') // ⚠️ Confirm field refactor here
-      .insert(payload) // ⚠️ Confirm field refactor here
-      .select(DRIVER_COLS)
-      .single();
-
-    if (error) {
-      handleSupabaseError('create driver', error);
-      throw error;
-    }
-
-    return data;
-  } catch (error) {
-    handleSupabaseError('create driver', error);
-    throw error;
-  }
-};
-
-export const updateDriver = async (id: string, updates: Partial<Driver>): Promise<Driver | null> => {
-  try {
-    // Remove undefined values
-    const cleanUpdates = Object.fromEntries( // ⚠️ Confirm field refactor here
-      Object.entries(updates).filter(([_, value]) => value !== undefined)
-    );
-
-    const { data, error } = await supabase // ⚠️ Confirm field refactor here
-      .from('drivers')
-      .update(cleanUpdates)
-      .eq('id', id)
-      .select(DRIVER_COLS)
-      .single();
-
-    if (error) {
-      handleSupabaseError('update driver', error);
-      throw error;
-    }
-
-    return data;
-  } catch (error) {
-    handleSupabaseError('update driver', error);
-    throw error;
-  }
-};
-
-export const deleteDriver = async (id: string): Promise<boolean> => {
-  const { error } = await supabase
-    .from('drivers')
-    .delete()
-    .eq('id', id);
-
-  if (error) {
-    handleSupabaseError('delete driver', error);
-    return false;
-  }
-
-  return true;
-};
-
 // Photo upload function for drivers
 export const uploadDriverPhoto = async (file: File, driverId: string): Promise<string | undefined> => {
   if (!file || !file.name) { // ⚠️ Confirm field refactor here
@@ -577,150 +57,6 @@ export const uploadDriverPhoto = async (file: File, driverId: string): Promise<s
     handleSupabaseError('upload driver photo', uploadError);
     throw uploadError;
   }
-};
-
-// Trip CRUD operations
-export const getTrips = async (): Promise<Trip[]> => {
-  try {
-    const { data: { user }, error: userError } = await supabase.auth.getUser(); // ⚠️ Confirm field refactor here
-    
-    if (userError) {
-      if (isNetworkError(userError)) {
-        if (config.isDev) console.warn('Network error fetching user for trips, returning empty array');
-        return [];
-      }
-      handleSupabaseError('get user for trips', userError);
-      return [];
-    }
-    
-    if (!user) {
-      console.error('No user authenticated');
-      return [];
-    }
-
-    const { data, error } = await supabase // ⚠️ Confirm field refactor here
-      .from('trips')
-      .select('id, vehicle_id, trip_start_date, trip_end_date, start_km, end_km, calculated_kmpl, driver_id, refueling_done, fuel_quantity, fuel_rate_per_liter, total_fuel_cost, warehouse_id')
-      .eq('added_by', user.id)
-      .order('trip_start_date', { ascending: false });
-
-    if (error) {
-      handleSupabaseError('fetch trips', error);
-      return [];
-    }
-
-    return data || [];
-  } catch (error) {
-    if (isNetworkError(error)) {
-      if (config.isDev) console.warn('Network error fetching user for trips, returning empty array');
-      return [];
-    }
-    handleSupabaseError('get user for trips', error);
-    return [];
-  }
-};
-
-export const getTrip = async (id: string): Promise<Trip | null> => {
-  const { data, error } = await supabase
-    .from('trips') // ⚠️ Confirm field refactor here
-    .select('*')
-    .eq('id', id)
-    .single();
-
-  if (error) {
-    handleSupabaseError('fetch trip', error);
-    return null;
-  }
-
-  return data;
-};
-
-export const createTrip = async (tripData: Omit<Trip, 'id'>): Promise<Trip | null> => {
-  try {
-    const userId = await getCurrentUserId();
-    if (!userId) {
-      throw new Error('User not authenticated');
-    }
-
-    // Sanitize empty strings to null for UUID fields to prevent Supabase validation errors
-    const sanitizedTripData = { ...tripData };
-    if (sanitizedTripData.fuel_station_id === '') {
-      sanitizedTripData.fuel_station_id = null;
-    }
-    if (sanitizedTripData.station === '') {
-      sanitizedTripData.station = null;
-    }
-    if (sanitizedTripData.vehicle_id === '') {
-      sanitizedTripData.vehicle_id = null;
-    }
-    if (sanitizedTripData.driver_id === '') {
-      sanitizedTripData.driver_id = null;
-    }
-    if (sanitizedTripData.warehouse_id === '') {
-      sanitizedTripData.warehouse_id = null;
-    }
-
-    const payload = withOwner(
-      {
-        ...sanitizedTripData,
-        station: sanitizedTripData.station ?? null,
-        fuel_station_id: sanitizedTripData.fuel_station_id ?? null,
-      },
-      userId
-    );
-
-    if (config.isDev) console.log("Submitting trip with payload:", payload);
-    const { data, error } = await supabase
-      .from('trips')
-      .insert(payload)
-      .select('*')
-      .single();
-
-    if (error) {
-      handleSupabaseError('create trip', error);
-      throw error;
-    }
-
-    return data;
-  } catch (error) {
-    handleSupabaseError('create trip', error);
-    throw error;
-  }
-};
-
-export const updateTrip = async (id: string, updates: Partial<Trip>): Promise<Trip | null> => {
-  try {
-    const { data, error } = await supabase
-      .from('trips')
-      .update(updates)
-      .eq('id', id)
-      .select('*')
-      .single();
-
-    if (error) {
-      handleSupabaseError('update trip', error);
-      throw error;
-    }
-
-    return data;
-  } catch (error) {
-    handleSupabaseError('update trip', error);
-    throw error;
-  }
-};
-
-export const deleteTrip = async (id: string): Promise<boolean> => {
-  const { error } = await supabase
-    .from('trips')
-    .delete()
-    .eq('id', id);
-
-  if (error) {
-    handleSupabaseError('delete trip', error);
-    return false;
-  }
-
-  return true;
 };
 
 // Warehouse operations
